chatbot/requirements.txt
Flask
google-generativeai
python-dotenv
requests
Flask-Session
pytz
SpeechRecognition

